refactor: 전역 변수 파일 통합 #20
* 여러 파일에서 참조되는 변수는 전역 파일 내로 추가
* 한 파일에서만 참조되는 변수는 지역 변수화
